fix(windit/plan,aggregate)!: a window coverage is a weight, so it is f64 end to end
`Span::coverage` returns `f64`, and `AggregatePolicy::aggregate_values` takes
`coverages: &[f64]`. Same defect class as the coefficient change this release
already carries, and this note names the class rather than the number: what
types a value is WHAT IT MULTIPLIES, not where it came from. A coverage was
typed by provenance — a window-geometry fraction rather than an embedding value,
therefore `f32` — while `CoverageWeightedMean` multiplied an embedding by it
inside an `f64` fold.

Two things that cost, and each has a falsifier written first and seen red.

OPERANDS ROUNDED BEFORE THE DIVISION. `f32` holds every integer only to `2^24`.
`WindowOptions::new(16_777_217)` over `16_777_216` elements plans one ragged tail
one element short of full; `16_777_217` is `2^24 + 1`, so the window narrowed to
`2^24` and the quotient came out exactly `1.0` — a ragged tail indistinguishable
from a full window, in a crate whose front page asserts `coverage() < 1.0` for
exactly that span. Red before:

  a tail one element short of the window must not report full coverage, got 1.0

THE QUOTIENT LANDED ON THE `f32` GRID, `2^-24` apart relatively where the fold
rounds at `2^-53`. Spans (len 8388607, window 16777213) and (len 8388608, window
16777215) — every operand at most `2^24` and so exactly representable, which is
what makes this about the division and not the operands — have true coverages
`1/(16777213*16777215)` apart: `3.6e-15`, 32 `f64` ulps, about `6e-8` of an `f32`
ulp. Both reached the fold as one weight. Red before:

  assertion `left != right` failed: spans (len 8388607, window 16777213) and
  (len 8388608, window 16777215) have true coverages 3.6e-15 apart, yet both
  folded at 0.50000006 / 0.50000006 and produced one vector
    left: [0.8944271696751094, 0.4472136381495684]
   right: [0.8944271696751094, 0.4472136381495684]

THE BREAK IS A DIFFERENT CLASS FROM THE COEFFICIENT'S. That one broke no
implementor. This changes an object-safe trait method's signature, so every
downstream `AggregatePolicy` stops compiling — one character, and the compiler
points at it (`_coverages: &[f32]` -> `&[f64]`). And it moves
`CoverageWeightedMean`'s output for every caller, with no source edit that opts
out, because the coverage is derived by the planner rather than configured. A
one-element tail in a window of three goes `0.3333333432674408` ->
`0.3333333333333333`, a relative `3.0e-8` — the eighth significant digit, the
same size as the coefficient change. Through a four-window fold:

  before  [0.9938837343123988, 0.11043152932582692]
  after   [0.993883734673619,  0.11043152607484655]

`3.6e-10` in the dominant component, `2.9e-8` in the one the tail steers. That is
a re-measure, not a recompile, and the changelog says so under its own heading.
`Span::coverage` and this weight are byte-identical between `0.2.0` and the
commit before this one, so `0.2.0` and "before" are the same baseline.

THE DESIGN QUESTION IS ANSWERED, NOT SIDESTEPPED. The old answer was documented
in three places, and `EmaRenormalized`'s note called it "a live question, not one
this note settles" while arguing for the shape from API cost. All three are
rewritten: the module note now says configuration is typed by what it multiplies
and answers the follow-up (why the slice is a concrete `f64` and not the policy's
`C` — a coefficient is CONFIGURED, so it can be asked for in the embedding's
domain; a coverage is DERIVED, by the planner from two `usize`s in a tier with no
compute scalar, so there is no `C` to ask it in). `Real::from_f32` is down to the
gate's own constant and `Real::from_f64` carries both weights.

ONE CONSEQUENCE IN THE INPUT DOMAIN, and it is not cosmetic. `[0, 1]` in `f64`
reaches `2^-1074` where `f32` stopped at `2^-149`, and that bound was doing quiet
work: it kept `CoverageWeightedMean`'s weight bounded below, so its products were
guaranteed normal. It no longer is, so the policy joins `EmaRenormalized` in the
regime the determinacy gate's `MIN_GATE_THRESHOLD` floor decides. `aggregate`
cannot reach it — `len >= 1` puts a plan-derived coverage at `1/usize::MAX` or
above — only a caller building its own slice can. The Input domain note now
distinguishes the policies the domain bounds below from those it does not, and a
test pins the boundary from both sides; mutating the floor to zero reds it with a
fabricated `[0.7071067811865475, 0.7071067811865475]`.

`cargo semver-checks` reports this no more than it reported the coefficient
break: 223 checks, 223 pass, "no semver update required", exit 0, forced as a
patch against `0.2.0`. That settles a question the last release left open — it
was reasonable to suppose a changed METHOD SIGNATURE on a public trait would be a
different story from a changed field type. It is not. Recorded in the changelog
so the tool's silence is never read as evidence again.

Mutation ledger re-run: 42 rows, 39 killed. Three survivors, each confirmed
SURVIVING AT `origin/main` TOO rather than assumed pre-existing —
`Ema::smoother`'s defensive re-clamp (the known one), the fold's Neumaier
compensation, and `l2_renorm`'s own zero rejection (shadowed by the determinacy
gate on every path that reaches it). The coverage-slice length check survived the
first pass and is now killed by a test.

+++ src/plan/mod.rs
@@ -106,9 +106,32 @@ impl Span {
   ///
   /// Both constructors enforce `0 < len <= window` in every build, so the
   /// denominator is positive and the quotient is finite.
+  ///
+  /// # The fraction is a weight, so it is `f64`
+  ///
+  /// This is not a report a caller merely reads: it is what the
+  /// `CoverageWeightedMean` aggregation policy multiplies an embedding by, in
+  /// the `f64` domain every shipped scalar computes in (named rather than
+  /// linked — that policy sits behind `alloc` and this tier is featureless). A
+  /// weight resolved more coarsely than the arithmetic it drives loses two
+  /// distinct things, and this fraction lost both while it was `f32`:
+  ///
+  /// - **The operands rounded before the division.** `f32` represents every
+  ///   integer only to `2^24`; `f64` to `2^53`. A window of `2^24 + 1` narrowed
+  ///   to `2^24`, so a tail one element short of it divided out to exactly `1.0`
+  ///   and a ragged tail was indistinguishable from a full window.
+  /// - **The quotient was rounded to the `f32` grid.** That grid is `2^-24`
+  ///   apart relatively where the fold rounds at `2^-53`, so two window
+  ///   geometries whose true coverages differ by less than an `f32` ulp — as
+  ///   `8388607/16777213` and `8388608/16777215` do, by `3.6e-15` — arrived at
+  ///   an `f64` fold as one weight.
+  ///
+  /// `f64` moves the first rounding out to `2^53` and the second to the fold's
+  /// own `2^-53`; a plan whose window exceeds `2^53` elements rounds again, and
+  /// there is no wider domain to move it to.
   #[must_use]
-  pub fn coverage(&self) -> f32 {
-    self.len as f32 / self.window as f32
+  pub fn coverage(&self) -> f64 {
+    self.len as f64 / self.window as f64
   }
 }
